+++ _tasks/generate_artifacts.ts
@@ -3,8 +3,7 @@ import * as $ from "../deps/scale.ts"
 import { emptyDir } from "../deps/std/fs.ts"
 import * as path from "../deps/std/path.ts"
 import dprintConfig from "../dprint.json" assert { type: "json" }
-import { devUser } from "../nets/mod.ts"
-import { compress } from "../util/compression.ts"
+import { compress, devUser } from "../util/mod.ts"
 
 export const DEV_USER_COUNT = 100_000
