+++ examples/dev/dev_users.eg.ts
@@ -6,7 +6,8 @@
  * funds. This simplifies signing extrinsics for submission to the given dev chain.
  */
 
-import { $, $sr25519, createDevUsers } from "capi"
+import { $, $sr25519 } from "capi"
+import { createDevUsers } from "capi/server"
 
 /// Dev users can be initialized with no count. The resulting collection is
 /// a record with 26 `Sr25519`s (one for every letter of the alphabet).
