+++ examples/misc/indices.eg.ts
@@ -7,8 +7,8 @@
 
 import { polkadotDev } from "@capi/polkadot-dev"
 import { assertEquals } from "asserts"
-import { createDevUsers } from "capi"
 import { signature } from "capi/patterns/signature/polkadot.ts"
+import { createDevUsers } from "capi/server"
 
 const { alexa } = await createDevUsers()
