+++ examples/sign/pjs.eg.ts
@@ -6,9 +6,10 @@
  */
 
 import { polkadotDev } from "@capi/polkadot-dev"
-import { createDevUsers, ss58 } from "capi"
+import { ss58 } from "capi"
 import { pjsSender, PjsSigner } from "capi/patterns/compat/pjs_sender.ts"
 import { signature } from "capi/patterns/signature/polkadot.ts"
+import { createDevUsers } from "capi/server"
 import { createPair } from "https://deno.land/x/polkadot@0.2.34/keyring/mod.ts"
 import { TypeRegistry } from "https://deno.land/x/polkadot@0.2.34/types/mod.ts"
