+++ examples/tx/handle_errors.eg.ts
@@ -7,8 +7,9 @@
 
 import { contractsDev } from "@capi/contracts-dev"
 import { assertInstanceOf } from "asserts"
-import { createDevUsers, ExtrinsicError } from "capi"
+import { ExtrinsicError } from "capi"
 import { signature } from "capi/patterns/signature/polkadot.ts"
+import { createDevUsers } from "capi/server"
 
 const { alexa, billy } = await createDevUsers()
