+++ examples/xcm/asset_teleportation.eg.ts
@@ -26,8 +26,9 @@ import {
   RuntimeEvent,
 } from "@capi/rococo-dev-westmint"
 import { assert } from "asserts"
-import { createDevUsers, Rune } from "capi"
+import { Rune } from "capi"
 import { signature } from "capi/patterns/signature/polkadot.ts"
+import { createDevUsers } from "capi/server"
 
 const { alexa } = await createDevUsers()
