+++ examples/xcm/reserve_transfer.eg.ts
@@ -27,9 +27,10 @@ import {
 } from "@capi/rococo-dev-xcm-statemine"
 import { rococoDevXcmTrappist } from "@capi/rococo-dev-xcm-trappist"
 import { assert, assertNotEquals } from "asserts"
-import { $, alice as root, createDevUsers, Rune, ValueRune } from "capi"
+import { $, alice as root, Rune, ValueRune } from "capi"
 import { $siblId } from "capi/patterns/para_id.ts"
 import { signature } from "capi/patterns/signature/statemint.ts"
+import { createDevUsers } from "capi/server"
 import { retry } from "../../deps/std/async.ts"
 
 const { alexa, billy } = await createDevUsers()
