+++ mod.ts
@@ -1,5 +1,6 @@
 export * as $ from "./deps/scale.ts"
 export { BitSequence } from "./deps/scale.ts"
+export { DevnetConnection } from "./server/client/DevnetConnection.ts" // TODO remove
 export * from "./util/autobin.ts"
 
 // moderate --exclude main.ts nets.ts server util
