+++ server/client/detectServer.ts
