+++ server/client/isUploaded.ts
@@ -1,0 +1,5 @@
+export async function isUploaded(server: string, hash: string) {
+  const url = new URL(`upload/codegen/${hash}`, server)
+  const exists = await fetch(url, { method: "HEAD" })
+  return exists.ok
+}
