+++ util/_artifacts/devUserNames.ts
@@ -1,0 +1,1 @@
+export const devUserNames = ["alexa", "billy", "carol", "david", "ellie", "felix", "grace", "harry", "india", "james", "kiera", "laura", "matty", "nadia", "oscar", "piper", "quinn", "ryann", "steff", "tisix", "usher", "vicky", "wendy", "xenia", "yetis", "zelda"] as const
