+++ util/mod.ts
@@ -4,6 +4,7 @@ export * from "./ArrayOfLength.ts"
 export * from "./autobin.ts"
 export * from "./clock.ts"
 export * from "./compression.ts"
+export * from "./devUser.ts"
 export * from "./error.ts"
 export * from "./gracefulExit.ts"
 export * from "./key.ts"
